Allow several trainer datasets
The idea is to keep old versions setups links valids with proper skill
positions and description.

Note that old urls are still valid

Add a `d` URL parameters for dataset version. If you're watching
a setup made with an older version of CoR a warning will be shown, as
well as a link leading you to the current trainer dataset version.

Add a fake 1.33.2 dataset for legacy urls.

Close #2