Add google analytics stuff
I don't especially like that kind of service but i want to see how much
the trainer is used.

Depending of the results, i may deploy an url shortener for shared URLs,
while keeping the long URLs for long term archival, meaning that
clicking on share will give you a long link for archival, and a short
one for sharing in the short term.