global: stop using jquery
Introduce lamaiquery, a tiny subset of jquery with all that is needed to
run the site:

- The size is reduced to 96% compared to jquery
- It's 20% faster when loading a setup in the trainer
- It has no silent exceptions, and as such uncovered some bad stuff,
  notably trainer.js trying to manipulate the WM row +/- spinners that
  don't exist

Unlike the rest of the site, it's MIT licensed so I can use it in a more
permissively licensed project later.

Data submission (collect/) has to be tested live due to CORS
restrictions, so there may be another commit after that one.