GLOBAL: use the fetch API instead of synchronous xmlhttprequest
This impacts mostly the trainer.

Synchronous xmlhttprequest is deprecated since years but is still here
and probably will never disappear. Fetch API syntax is way easier to
deal with and has nice error handling (even ithough dealing with Promises is
scary), and is the way to go currently. So why I didn't use it back in
the day?

Synchronous requests were needed in the trainer because when you load a
saved setup, actually your browser clicks for you, which validates your
saved setup in the same time. Which is nice as it shortens the code by a
lot (no double check to make), but as such the interface needs to be
loaded from trainer datasets BEFORE doing the automated clicks, which
isn't granted using the fetch API even in a "synchronous" context (aka
async/await).

So on top of that API change, loading a setup from a saved URL has been
split in two functions:

- The first one decompresses the setup and clicks on your
  level/class/version then finally on "load" (load_setup_from_url)
- The second one is called by load_tree() once the tree is effectively
  fully loaded, and the callbacks put in place (input_saved_setup_from_url)

As usual, the setup collection has to be tested live, so some further
commits may happen. And waiting for people to whine if stuff happens.