TRAINER: fix race condition on VERY slow connections
If you are using a slow connection (GPRS/2G), it was possible to click
on load while loading the trainer data; leading to load several time the
trainer, which not only display your setup, but several more empty setups,
giving you an already complete setup but with all your discipline and
power points available.

Hide the whole setup form while loading the tree, hoping it will be
enough.