Add WEVENTS: the full list of all events but relic ones ...
... in the last 30 days. It has infinite scrolling and various filters
to reduce the number of events. Given the size of the generated JSON,
that i estimate at 500kb when fully loaded, if you want to serve it's
recommended to use gzip/brotli compression in your webserver which
reduce the size to a few kilobytes at level 4 only!

It's generated through the stats.generate python module as it's a simple
SELECT query. Right now it's generated as needed, but i may throttle it
in the future.