GLOBAL: improve Cumulative Layout Shift
I would not like users having too much motion sickness right? Note that
people on width > 1600 will barely notice this.

Only events have to keep a big CLS, due to the random length of the
results and paginate this will be a pain in the ass to code and for
users to navigate.