GLOBAL: downgrade javascript to ES8 (>= 2017) instead of ES10 (>=2019)
Done by just moving to old school catch syntax with unused variables.

The site is now totally usable with Firefox 52.9 on Win XP with some
visual glitches. It's not planned to fix these glitches, as this is
extra bloat for virtually no users (would be fun though).

Also add contributions guidelines with current baseline support. It's still
missing a handler for non ES8 browsers that would indicate that the
site is simply unusable, because the use of async functions; it would be
needed to do funky things to make it work according to the ES version
supported. I'll add one after that.