GLOBAL: add hit counter
This only works for mascaldotfr.github.io and pings the API server to
generate basic stats. Putting it in browserquirks is pretty meh but this
avoids yet another fetch. This may be removed if the data is not
relevant.

The goal is to track the bandwidth usage of the site (and HTTP 304
efficiency), browsers used and visitors countries.

Only 1 week of data is kept and just tracks hits. Some referers have
been anonymised, because you already hit my server if you don't serve
`warstatus/` yourself and i don't want to expose people when they
wouldn't want to.

The data is available at https://hail.thebus.top/cortdata_stats.html
FYI.