TRAINER: add a hidden beta page for Amun
While it would be possible to add beta versions to the main trainer,
they would sit on top of the selector until the new live version is out,
and bring clutter, which is not much wanted, especially that beta and rc
are anyway very short lived (see the discord amun changelogs about rc
and beta delay).

This new page will hold a single "beta" version of the skilltrees,
allowing to update the trainer dataset incrementally at every beta and
rc, and possibly to bring the live version change with 0-day delay.

Converting from beta to latest live is possible through a magic link.

I hope github page handles symlinks well.