mynotify: asking permissions is now more civilised
Introduce a bell icon to subscribe to notifications, instead of throwing
the permissions dialog right in our face. The bell vanishes if a choice
has been made. If the permissions choice is reset, then the bell appear
again.

Prompted by Firefox emitting errors, as asking permissions should be
initiated by the user.