TRAINER: drop most use of string.concat and a few useless template literals
string.concat is slow but powerful (it magically converts everything to
string), but useless in most case in the trainer due to the use of
templates string on top of that. I even found out that we stringify
stringified strings!

While here remove some useless use of template literals.

No functional changes, as i don't think someone is lagging while show a
skill description or sharing a setup :P