WEVENTS: record relic events from now on
We can expect a change in the invasions system, the question is when ...
That will give me more time to focus on other things when it does, and
lead to almost no overhead as anyway they're not used to generate stats
at all.

Note that as I'm committing this, no relic event is recorded yet, it has
been tested through a modified official warstatus page.