TRAINER: performances fixes
There was and still is a lot of naivety in the trainer code, some of
them has been fixed. Notably a skeleton has been added to reduce CLS and
now trainer rows redrawing is almost instant since the redraw occurs
after the trainerdata has been downloaded. It's way WAY smoother than it
used to be.

That last point cause issues while changing classes in test usage, so a
fake lag (ahahah) through a CSS animation playing with opacity has been
added. Should be OK even for people with motion sickness.