BZ/BOSSES: fix a race condition if visited for the frist time
It appears that if you visit the page for the first time, language or tz
in localStorage is null when date and time formatters are created,
leading to a blank page due to a race condition.

Move all these thing in document.ready().

The bug has been introduced by 48f7681 a week ago, and some peeps
probably met it without reporting. Sad.

Survived a few private browsing sessions, that I should have used
earlier. Maybe I should assert tz and lang, maybe ...to be continued.