TRAINER: fix bad_shared_link()
I've bikeshedded the thing because such occurrence is very rare, but
Chrome (and all its variants) loop indefinitely if a shared setup link
is bad, because it doesn't like :

```
window.alert(...)
window.location.replace(...)
```

Fixed it by replacing the trainer body by an alert message and countdown
for redirection to the trainer.

Also, fixed some bad calls to the function, and removed the legacy
parameter for this function.

You can visit it from that link:
https://mascaldotfr.github.io/CoRT/?t=aea8baaaaabaaaaabaaaaabaaaaabaaaaabaaaaaZaaaaa