BOSSES: move from png to webp
webp support: https://caniuse.com/webp

Images size are divided by 3.

It wasn't done before because Safari adopted the format lately (2022).

Actually, it's not a problem for old desktop Macs where you can use
another browser, but for iOS, it was an issue since they force using
Webkit even with when using Firefox or Chrome. But it's no more 2022
anymore.

That change still allows old iPhones compatible with iOS 14 to use CoRT
as usual, meaning iPhones 6s (released in _2015_) and later are
supported.