TRAINER: yet another load of performance improvements
This mostly concerns saved setups, which used to be slow to load.

Loading the whole tree for a full lvl 60 conjurer saved setup is now 10
times faster than in last June, and 3 times faster than in the previous
commit. Also the global page loading is 2.5 times faster.

In this commit, we're still reducing useless DOM reads and writes:

- Cache selectors when doing multiple click on the same power /
  discipline. And stop clicking, we call directly functions instead.
- Instead of hiding '+' and '-' buttons individually, hide the whole
  spinner instead
- "data-value" attr has become useless for power points (PP) and
  discipline points (DP) left
- Skip intermediate UI updates during automated setup loading,
  `input_from_url()` now only updates trees' visual state once by tree,
  and set PP and DP left only once the whole setup has been put in
  place.
- Defer tooltips binding so skillbars are shown earlier
- Use querySelector for fetching data from the DOM whenever it's
  possible, instead of getElementsByClassName().

Survived tests.html. Special thanks to Chrome's profiler <3