GLOBAL: add an API endpoint for maintenance
Until now, I had to commit something to add maintenance messages. It was
unpractical and prone to issues with cache. And it does not work anymore
due to using stable releases for github.

So we add an API endpoint that, if not empty, will display the content
of what the endpoint returns (see the docs at `/api/MAINTENANCE.md`) in
a big dark purple (maintenance) or red (API server unreachable) card on
top of pages.

It's fetched a few hundreds milliseconds after the page is loaded, so it
has no impact on performance.