deploy: docker changes, and some rewording here and there
Some notes:

Not a big fan of Docker personnally, but it's really slick to test
things from scratch. The problem is that a potential user of said docker
environment want very likely a prepopulated container.

So we prepopulate as default, and if not needed, setting the
`PREPOPULATE` environment variable to `no` during build time won't
prepopulate API data.

Also, the resulting container don't use "cort" as default user but
"www-data", to simplify things.

wget(1) calls in various places have been streamlined as well, so we
don't call it for every URL