WZ: reduce server calls (take 2)
Previous attempt had 2 issues:

1. Slow setInterval() drifting. It is as it is, if the tab fall in the
   background category it drifts
2. On mobile, if the device slept, the timer drifted and refused to be
   back in the right second.

So this has been implemented instead:

1. For stable setups (desktop pc), ensure that the timer is never
   drifting, by resetting the timer starting second, if it appears
   than the time elapsed since the last run is > 60s.
2. For unstable situations (mobile), on focus, if we find out that the
   last API fetch was > 60s, then reschedule by reinstanciating. That
   allows to update immediatly on device wake up or if the tab has been
   put in sleep mode and suddendly woken up. Nice bucket of water :)

That seems aggressive, especially on mobile, but operating like this
reduces the number of API calls a lot, with minimal overhead.

To be more "field tested" but it should be fine... and if it is, I'll
add this to Sentinel later.

Close #26 (for now?)