collect: stop rewriting all saved setups everytime
This has been the first piece of server side stuff I wrote for CoRT 3
years ago. Back then I didn't expect to see CoRT becoming that popular
and implemented a full file rewrite for every submitted setup as quick
hack (and well my PHP was rusty, and still is).

With almost 10K setups nowadays, it's beginning to get expensive, so
only add the line corresponding to the setup to the
trainer_saved_setups.txt and its gziped counterpart.

The submission runtime is now around 1.5ms; it used to be 35ms before
that commit.