GLOBAL: generate less html tags from javascript
This looks a big change, but it isn't. tstats.html is barely touched
because it is the standard used for all the other pages.

This change targets two kind of people :

1. People on very low end devices, who will quickly notice the
   performance boost, especially for WZ, where it's 2 times faster
   according to profiling at throttle x20, due to dynamically loading
   some complex layout and array manipulations removed.
2. People working on CoRT codebase, as that approach is cleaner.

While here, let lamaiquery's element.text() really set up the text,
before that text() was an alias of html(), shaving a few ms here and
there.

The menu has been sparred, as it is a lot of duplication for no real
gain. I'll try a different approach later.