BOSSES: use Regnum Tools drifting offsets
Regnum tools manage to get fairly accurate respawns times despite not
having been updated since ages (Well, it misses TK's weird leap). So
it's interesting to me to know how it has been done.

It appears that the drift is NOT equal for each boss. Surprising. Here
are the values :

Eve: 7s
Daen: 5s
TK: 3s

Won't put a hotfix for that, it's a 1 minute issue until next release.