BZ/BOSSES: drop atcb in favor of simple ICS generation
This library is awesome to be honest. But I only use a tiny part of it,
and it's 500kb ungzipped, bigger than CoRT actually (minus the trainer
data).

It's 2025, people have smartphone, Windows has a calendar app that
understands ICS out of the box. Just providing the ICS is good enough,
especially that almost no one use this feature, but at least I need it.

So we implement an easy to use calendar link generator. It uses unix
timestamps since API data returns that. It's only 2.5k ungzipped. Nice.

While here add 1H and 10M reminders by default.