BZ: fix display in exotic locations for the player base
BZ were misplaced in place far from UTC0 (LA, Sydney, Tokyo). This
commit fixes this, by calculating each future BZ date and placing in the
corresponding day. Further magic was need for negative UTC offset.

Javascript in browser is pretty terrible for this, since you can set up
a singletimezone for the whole script.

CoRT's timezone is ignored, and it's marked as such in the page. Too
much of a hassle already.