BZ: refactor schedule
Rewrite the scheduler from scratch, leading to a major unbloat and
simplification. I should not have rushed the feature with parts of AI
slops. Not that the camel case in it is voluntary, it is as it was the
original function.

The current day will always be the first row.

Tested with LA, Paris and Sydney.